KernelCpuUidTimeReader: Always check if `remove_uid_range` file exists
* remove_uid_range doesnt exist on some kernels, always check if the file exists first to avoid causing logspam and IOException

+++ core/java/com/android/internal/os/KernelCpuUidTimeReader.java
@@ -31,6 +31,7 @@ import com.android.internal.os.KernelCpuProcStringReader.ProcFileIterator;
 import com.android.internal.os.KernelCpuUidBpfMapReader.BpfMapIterator;
 
 import java.io.BufferedReader;
+import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.nio.CharBuffer;
@@ -301,6 +302,11 @@ public abstract class KernelCpuUidTimeReader<T> {
             removeUidsFromKernelModule(startUid, endUid);
         }
 
+         public boolean fileExists(String fileName) {
+             final File file = new File(fileName);
+            return file.exists();
+         }
+    
         /**
          * Removes UIDs in a given range from the kernel module and internal accounting data. Only
          * {@link BatteryStatsImpl} and its child processes should call this, as the change on
@@ -311,6 +317,7 @@ public abstract class KernelCpuUidTimeReader<T> {
          * @param endUid   the last uid to remove
          */
         private void removeUidsFromKernelModule(int startUid, int endUid) {
+            if (!fileExists(REMOVE_UID_PROC_FILE)) return;
             Slog.d(mTag, "Removing uids " + startUid + "-" + endUid);
             final int oldMask = StrictMode.allowThreadDiskWritesMask();
             try (FileWriter writer = new FileWriter(REMOVE_UID_PROC_FILE)) {
